📦 Feat (ArticuloInsumo): Implementar DTOs y mappers para ArticuloInsumo, actualizar ArticuloInsumoController y ArticuloInsumoService para usar DTOs, ajustar métodos de creación, actualización y eliminación, corregir tipo de precioVenta en ArticuloDTO y ArticuloInsumoDTO.

+++ src/main/java/utn/saborcito/El_saborcito_back/dto/ArticuloDTO.java
@@ -10,7 +10,7 @@ import lombok.AllArgsConstructor;
 public class ArticuloDTO {
     private Long id;
     private String denominacion;
-    private Integer precioVenta;
+    private Double precioVenta; // Corregido de Integer a Double
     private Long categoriaId;
     private Long unidadMedidaId;
 }
